+++ src/settings/sets/gallery.js
@@ -5,6 +5,7 @@ export const GallerySet = {
     searchEvent: 'gallery-search-open',
     searchSavedInit: 'gallery-saved-init',
     searchSave: true,
+    searchEnabled: true,
     thumbsOpenable: '1',
     minimumSearch: 3, // chars
     perPage: 4, // all - without pagination