+++ src/components/ui/generalNote/theme.js
@@ -1,0 +1,7 @@
+export const generalNoteTheme = {
+    common: {
+       success: 'green',
+       warning: 'orange',
+       error: 'red'    
+    }
+};