+++ src/components/ui/paginatableContent/theme.js
@@ -1,0 +1,17 @@
+export const paginatableContentTheme = {
+    common: {
+       border: 'grey',
+       background: 'white',
+       active: '#dcdcdc'    
+    },
+    main1: {
+        border: 'grey',
+        background: 'white',
+        active: '#dcdcdc'  
+    },
+    blue: {
+        border: 'grey',
+        background: 'white',
+        active: '#96cfeb'  
+    }
+};