+++ src/enums/arrays.js
@@ -1,0 +1,4 @@
+export const ArrayEnums = {
+    All: 'all',
+    None: 'none'
+};