+++ src/enums/index.js
@@ -1,3 +1,5 @@
 
 export { KeyboardKeys } from './keyboard';
 export { GeneralNoteEnums, GeneralNoteCodes } from './generalNote';
+export { Cursors } from './styles';
+export { ArrayEnums } from './arrays';