+++ src/enums/styles.js
@@ -1,0 +1,8 @@
+export const Cursors = {
+    normal: 'auto',
+    pointer: 'pointer',
+    zoomIn: 'zoom-in',
+    zoomOut: 'zoom-out',
+    wait: 'wait',
+    notAllowed: 'not-allowed'
+};