+++ src/pages/additional/gallery/theme.js
@@ -1,0 +1,6 @@
+export const GalleryPageTheme = {
+    common: {
+        background: '#f2f2f2',
+        border: 'grey'
+    }
+};