+++ src/pages/additional/index.js
@@ -1,6 +1,6 @@
 import './game';
 import './writer';
-import './gallery';
+import './gallery/gallery';
 import './map';
 import './tabs';
 import './account';