+++ src/theme/red.js
@@ -1,3 +1,7 @@
+import { paginatableContentTheme } from '../components/ui/paginatableContent/theme';
+import { generalNoteTheme } from '../components/ui/generalNote/theme';
+import { GalleryPageTheme } from '../pages/additional/gallery/theme';
+
 const mainColor = '#ad0e2b';
 
 export const themeRed = {
@@ -41,11 +45,9 @@ export const themeRed = {
             }
         }
     },
-    generalNote: {
-        success: 'green',
-        warning: 'orange',
-        error: 'red'
-    },
+    generalNote: generalNoteTheme.common,
+    paginatableContent: paginatableContentTheme.common,
+    galleryPage: GalleryPageTheme.common,
     settings: {
         textSize: {
             border: '#cf3030',