+++ src/settings/sets/imageViewer.js
@@ -1,6 +1,7 @@
 export const ImageViewerIds = {
     common: 'common',
-    writer: 'writer'
+    writer: 'writer',
+    gallery: 'gallery'
 };
 
 export const ImageViewerSettings = {
@@ -8,7 +9,7 @@ export const ImageViewerSettings = {
     allLevelsCount: true,
     propertyAmount: 9,
     common: { // default, non-specific
-        // searchListNum: 0 - fist item, 1 - 2 items, [num] - integer items, 'all' - all items
+        // searchListNum: 0 - fist item, 1 - 2 items, [num] - integer items, ArrayEnums.All - all items
         searchListNum: 0,
         draggable: true,
         zoomEnable: true,
@@ -21,6 +22,18 @@ export const ImageViewerSettings = {
         }
     },
     writer: { // writer - id
+        searchListNum: 0,
+        draggable: true,
+        zoomEnable: true,
+        originalLink: true,
+        exceptionHandler: true,
+        zoom: {
+            min: 0.4,
+            max: 3,
+            keyboard: true
+        }
+    },
+    gallery: {
         searchListNum: 0,
         draggable: true,
         zoomEnable: true,