+++ Changelog.md
@@ -1,6 +1,7 @@
 ## Changelog
 v 2.32
-  Added Gallery updates, Added GallerySearch separate component
+  Added Gallery updates, Added GallerySearch separate component,
+  First time use UI Component mountable theme
   
 v 2.30
   Created Gallery First version, added theme GeneralNote,