+++ Changelog.md
@@ -1,7 +1,8 @@
 ## Changelog
 v 2.32
   Added Gallery updates, Added GallerySearch separate component,
-  First time use UI Component mountable theme
+  First time use UI Component mountable theme,
+  Added ImageViewer to open every Gallery image
   
 v 2.30
   Created Gallery First version, added theme GeneralNote,